16 bit hash

+++ src/engine/transposition_table.h
@@ -16,9 +16,9 @@ enum {
     upper_bound
 };
 
-typedef uint64_t TTLength_t;
+typedef uint32_t TTLength_t;
 typedef uint32_t TTIndex_t;
-typedef uint64_t TTKey_t;
+typedef uint16_t TTKey_t;
 
 typedef struct {
     TTFlag_t flag;
